Move exception to the same path as master branch
Fix CS from Dummy docblock
